Refactored items into specific subfolders for better code management
