+++ client/shell/src/app/shared/_components/tangy-svg-logo/tangy-svg-logo.component.ts
