+++ client/shell/src/app/shared/_directives/seamless-with-window.directive.spec.ts
