+++ client/shell/src/app/shared/_pipes/truncate-value.pipe.spec.ts
